Making GPU interpolation great again!
