+++ learn2track/models/gru.py
@@ -149,127 +149,3 @@ class GRU(Model):
         model = cls(**hyperparams)
         model.load(path)
         return model
-
-
-class GRU_Regression(GRU):
-    """ A standard GRU model with a regression layer stacked on top of it.
-    """
-    def __init__(self, input_size, hidden_sizes, output_size, **_):
-        """
-        Parameters
-        ----------
-        input_size : int
-            Number of units each element Xi in the input sequence X has.
-        hidden_sizes : int, list of int
-            Number of hidden units each GRU should have.
-        output_size : int
-            Number of units the regression layer should have.
-        """
-        super().__init__(input_size, hidden_sizes)
-        self.output_size = output_size
-        self.layer_regression = LayerRegression(self.hidden_sizes[-1], self.output_size)
-
-    def initialize(self, weights_initializer=initer.UniformInitializer(1234)):
-        super().initialize(weights_initializer)
-        self.layer_regression.initialize(weights_initializer)
-
-    @property
-    def hyperparameters(self):
-        hyperparameters = super().hyperparameters
-        hyperparameters['output_size'] = self.output_size
-        return hyperparameters
-
-    @property
-    def parameters(self):
-        return super().parameters + self.layer_regression.parameters
-
-    def _fprop(self, Xi, *args):
-        outputs = super()._fprop(Xi, *args)
-        last_layer_h = outputs[len(self.hidden_sizes)-1]
-        regression_out = self.layer_regression.fprop(last_layer_h)
-        return outputs + (regression_out,)
-
-    def get_output(self, X):
-        outputs_info_h = []
-        for hidden_size in self.hidden_sizes:
-            outputs_info_h.append(T.zeros((X.shape[0], hidden_size)))
-
-        results, updates = theano.scan(fn=self._fprop,
-                                       outputs_info=outputs_info_h + [None],
-                                       sequences=[T.transpose(X, axes=(1, 0, 2))])  # We want to scan over sequence elements, not the examples.
-
-        self.graph_updates = updates
-        # Put back the examples so they are in the first dimension.
-        self.regression_out = T.transpose(results[-1], axes=(1, 0, 2))
-        return self.regression_out
-
-    def use(self, X):
-        directions = self.get_output(X)
-        return directions
-
-
-class GRU_RegressionAndBinaryClassification(GRU):
-    """ A standard GRU model with a regression layer stacked on top of it
-    combined with a fully connected layer with a sigmoid non-linearity
-    to learn when to stop.
-    """
-    def __init__(self, input_size, hidden_sizes, output_size, **_):
-        """
-        Parameters
-        ----------
-        input_size : int
-            Number of units each element Xi in the input sequence X has.
-        hidden_sizes : int, list of int
-            Number of hidden units each GRU should have.
-        output_size : int
-            Number of units the regression layer should have.
-        """
-        super().__init__(input_size, hidden_sizes)
-        self.output_size = output_size
-        self.layer_regression = LayerRegression(self.hidden_sizes[-1], self.output_size)
-        self.stopping_layer = LayerDense(self.hidden_sizes[-1] + input_size, 1, activation="sigmoid", name="stopping")
-
-    def initialize(self, weights_initializer=initer.UniformInitializer(1234)):
-        super().initialize(weights_initializer)
-        self.layer_regression.initialize(weights_initializer)
-        self.stopping_layer.initialize(weights_initializer)
-
-    @property
-    def hyperparameters(self):
-        hyperparameters = super().hyperparameters
-        hyperparameters['output_size'] = self.output_size
-        return hyperparameters
-
-    @property
-    def parameters(self):
-        return super().parameters + self.layer_regression.parameters + self.stopping_layer.parameters
-
-    def _fprop(self, Xi, Xi_plus1, *args):
-        outputs = super()._fprop(Xi, *args)
-        last_layer_h = outputs[len(self.hidden_sizes)-1]
-        regression_out = self.layer_regression.fprop(last_layer_h)
-        stopping = self.stopping_layer.fprop(T.concatenate([last_layer_h, Xi_plus1], axis=1))
-
-        return outputs + (regression_out, stopping)
-
-    def get_output(self, X):
-        outputs_info_h = []
-        for hidden_size in self.hidden_sizes:
-            outputs_info_h.append(T.zeros((X.shape[0], hidden_size)))
-
-        results, updates = theano.scan(fn=self._fprop,
-                                       outputs_info=outputs_info_h + [None, None],
-                                       sequences=[{"input": T.transpose(X, axes=(1, 0, 2)),  # We want to scan over sequence elements, not the examples.
-                                                   "taps": [0, 1]}],
-                                       n_steps=X.shape[1]-1)
-
-        self.graph_updates = updates
-        # Put back the examples so they are in the first dimension.
-        self.regression_out = T.transpose(results[-2], axes=(1, 0, 2))
-        self.stopping = T.transpose(results[-1], axes=(1, 0, 2))
-
-        return self.regression_out, self.stopping
-
-    def use(self, X):
-        directions = self.get_output(X)
-        return directions
