+++ scripts/learn.py
@@ -9,6 +9,7 @@ sys.path = [os.path.abspath(os.path.join(__file__, '..', '..'))] + sys.path
 
 import numpy as np
 import argparse
+from os.path import join as pjoin
 
 import theano.tensor as T
 
@@ -24,7 +25,8 @@ from learn2track.factories import optimizer_factory
 from learn2track.factories import model_factory
 from learn2track.factories import loss_factory
 
-from learn2track.batch_schedulers import StreamlinesBatchSchedulerMultiSubjects
+from learn2track import datasets
+from learn2track.batch_schedulers import TractographyBatchScheduler
 
 
 def build_train_gru_argparser(subparser):
@@ -60,14 +62,9 @@ def build_argparser():
     # Dataset options
     dataset = p.add_argument_group("Data options")
     dataset.add_argument('--train-subjects', nargs='+', required=True,
-                         help='folder(s) containing `dwi.nii.gz` and `dataset.npz` that will be used as training data.')
+                         help='file containing training data (as generated by `process_streamlines.py`).')
     dataset.add_argument('--valid-subjects', nargs='+', required=True,
-                         help='folder(s) containing `dwi.nii.gz` and `dataset.npz` that will be used as validation data.')
-
-    dataset.add_argument('--dwi-name', default='dwi.nii.gz',
-                         help='name of the dwi file to look for in each subject folder (.nii|.nii.gz). Default: `%(default)s`')
-    dataset.add_argument('--dataset-name', default='dataset.npz',
-                         help='name of the dataset file to look for in each subject folder (.npz). Default: `%(default)s`')
+                         help='file containing validation data (as generated by `process_streamlines.py`).')
 
     duration = p.add_argument_group("Training duration options")
     duration.add_argument('--max-epoch', type=int, metavar='N', default=100,
@@ -121,17 +118,22 @@ def main():
 
     hyperparams_to_exclude = ['max_epoch', 'force', 'name', 'view']
     experiment_path, hyperparams, resuming = utils.maybe_create_experiment_folder(args, exclude=hyperparams_to_exclude)
+
+    # Log the command currently running.
+    with open(pjoin(experiment_path, 'cmd.txt'), 'a') as f:
+        f.write(" ".join(sys.argv) + "\n")
+
     print("Resuming:" if resuming else "Creating:", experiment_path)
 
     with Timer("Loading dataset", newline=True):
-        trainset = utils.load_streamlines_datasets(args.train_subjects, args.dwi_name, args.dataset_name)
-        validset = utils.load_streamlines_datasets(args.valid_subjects, args.dwi_name, args.dataset_name)
-        print("Datasets:", len(trainset), " |", len(validset))
+        trainset = datasets.load_tractography_dataset(args.train_subjects)
+        validset = datasets.load_tractography_dataset(args.valid_subjects)
+        print("Dataset sizes:", len(trainset), " |", len(validset))
 
-        batch_scheduler = StreamlinesBatchSchedulerMultiSubjects(trainset,
-                                                                 batch_size=args.batch_size,
-                                                                 noisy_streamlines_sigma=args.noisy_streamlines_sigma,
-                                                                 seed=args.seed)
+        batch_scheduler = TractographyBatchScheduler(trainset,
+                                                     batch_size=args.batch_size,
+                                                     noisy_streamlines_sigma=args.noisy_streamlines_sigma,
+                                                     seed=args.seed)
         print ("An epoch will be composed of {} updates.".format(batch_scheduler.nb_updates_per_epoch))
         print (batch_scheduler.input_size, args.hidden_sizes, batch_scheduler.target_size)
 
@@ -184,10 +186,10 @@ def main():
         # trainer.append_task(tasks.Print("Trainset - Error        : {0:.2f} | {1:.2f}", train_error.sum, train_error.mean))
 
         valid_loss = loss_factory(hyperparams, model, validset)
-        valid_batch_scheduler = StreamlinesBatchSchedulerMultiSubjects(validset,
-                                                                       batch_size=1000,
-                                                                       noisy_streamlines_sigma=None,
-                                                                       seed=1234)
+        valid_batch_scheduler = TractographyBatchScheduler(validset,
+                                                           batch_size=1000,
+                                                           noisy_streamlines_sigma=None,
+                                                           seed=1234)
 
         valid_error = views.LossView(loss=valid_loss, batch_scheduler=valid_batch_scheduler)
         trainer.append_task(tasks.Print("Validset - Error        : {0:.2f} | {1:.2f}", valid_error.sum, valid_error.mean))
